services/python/logAnalyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-multipart==0.0.6
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1 